Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/clients-matchall-exact-controller.https-expected.txt



FAIL Test Clients.matchAll() with exact controller assert_array_equals: property 1, expected false but got true
